backend/rag_service_api/pyproject.toml
[project]
name = "saviant-rag-micro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.13",
    "fastapi>=0.115.14",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.26",
    "langchain-google-genai>=2.1.5",
    "langchain-groq>=0.3.4",
    "langchain-huggingface>=0.3.0",
    "langchain-tavily>=0.2.4",
    "langgraph>=0.5.0",
    "langsmith>=0.4.4",
    "openai>=1.93.0",
    "pandas>=2.3.0",
    "pymupdf>=1.26.1",
    "pypdf>=5.6.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-socketio>=5.13.0",
    "pyyaml>=6.0.2",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.46.1",
    "uvicorn>=0.34.3",
]
